fix(memory): switch memory_read to int and cast callers to uint8_t
`memory_read()` previously returned `uint8_t` but attempted to signal errors by
returning negative errno values (e.g. -EINVAL). Under -Wsign-conversion with
-Werror, this produced a hard failure due to signed-to-unsigned wrap:

  unsigned conversion from `int` to `uint8_t` changes value from `-22` to `234`
  [-Werror=sign-conversion]

Change `memory_read()` to return `int` (errno-style) so failures can be
represented without implicit wrapping. Update call sites that expect a byte
value (opcode fetch and FX65 load) to explicitly cast the returned value to
`uint8_t`, preserving existing semantics while keeping strict builds clean.

+++ inc/memory.h
@@ -12,7 +12,7 @@ typedef struct __attribute__((packed)) {
 } Memory;
 
 void memory_init(Memory *m);
-uint8_t memory_read(const Memory *m, uint16_t addr);
+int memory_read(const Memory *m, uint16_t addr);
 int memory_write(Memory *m, uint16_t addr, uint8_t val);
 int memory_load_rom(Memory *m, const char *path);
